Major build system overhaul
* The common dbpp lib and adapters now are in different export sets.
* The adapters are components of the dbpp package. Consumers of this library
  should run find_package(dbpp COMPONENTS adapter1 adapter2 ..)
* Introduced three new build variants - Dev, DevWithCoverage and DevWithSanitizers
  The dev modes enable strict warnings, and tests are enabled by default
* Default symbol visibility is hidden also for non-windows platfors
* All installed files belong to different components, which is useful for packaging

+++ test/TestResult.cpp
@@ -142,7 +142,7 @@ TEST_CASE("Result", "[api]") {
                 ")");
 
         const int intVal = 14;
-        const float realVal = 3.14;
+        const float realVal = 3.14f;
         const std::string strVal = "/tmp/a/string/thats/also/a/path";
         std::vector<unsigned char> blobVal(1024);
         std::iota(blobVal.begin(), blobVal.end(), 0);
@@ -269,7 +269,7 @@ TEST_CASE("Result", "[api]") {
             REQUIRE(realRes.get(0, floatOut) == true);
             REQUIRE(floatOut == realVal);
 
-            float doubleOut = realVal * 2;
+            double doubleOut = realVal * 2;
             REQUIRE(realRes.get(0, doubleOut) == true);
             REQUIRE(doubleOut == realVal);
